Patch runtime's internal dynamic_import to apply asset base URL as needed (#6)
This was needed because when the .NET WebAssemblyHost starts with hot reload support enabled, it fetches the hot reload JS script from an absolute URL path (/_framework/hot-reload.js). Prior to .NET 8, this import was handled by code bundled in the Blazor startup script, so it was possible to apply the asset base URL there. Since .NET 8, it's now handled by the .NET browser runtime's internal dynamic_import function. This is available on the internal API object Blazor receives when it creates the runtime object. Since we can't modify the runtime's implementation of dynamic_import, we instead define an interceptor that will apply the asset base URL to imports that use absolute or relative URL paths.